Files & Threads

+++ Server/src/main/java/pt/isec/pd_g33/server/file/ThreadReceiveFiles.java
@@ -22,34 +22,38 @@ public class ThreadReceiveFiles implements Runnable {
 
     @Override
     public void run() {
-            try {
-                System.out.println("\nSOU O PORTO DO RECEIVE: Ip: " + ip + "Porto: " + port);
-                Socket socket = new Socket(ip, port);
-                InputStream in = socket.getInputStream();
 
-                ObjectOutputStream out = new ObjectOutputStream(socket.getOutputStream());
-                out.writeUnshared(filename);
+        if(new File(path + "/" + filename).exists())
+            return;
 
-                FileOutputStream fileOutputStream = new FileOutputStream(Paths.get(path + filename).toString());
-                byte[] buf = new byte[DATA_SIZE];
-                while(true) {
-                    int nBytes = in.read(buf);
-                    if(nBytes == -1)
-                        break;
-                    System.out.println("Li: " + nBytes);
-                    fileOutputStream.write(buf, 0, nBytes);
-                    buf = new byte[DATA_SIZE];
-                }
+        try {
+            System.out.println("\nSOU O PORTO DO RECEIVE: Ip: " + ip + "Porto: " + port);
+            Socket socket = new Socket(ip, port);
+            InputStream in = socket.getInputStream();
 
-                System.out.println("Received file '" + filename + "' from " + socket.getInetAddress().getHostAddress() + ":" + socket.getPort());
+            ObjectOutputStream out = new ObjectOutputStream(socket.getOutputStream());
+            out.writeUnshared(filename);
 
-                fileOutputStream.close();
-                in.close();
-                out.close();
-                socket.close();
-            } catch (IOException e) {
-                e.printStackTrace();
+            FileOutputStream fileOutputStream = new FileOutputStream(Paths.get(path + filename).toString());
+            byte[] buf = new byte[DATA_SIZE];
+            while(true) {
+                int nBytes = in.read(buf);
+                if(nBytes == -1)
+                    break;
+                System.out.println("Li: " + nBytes);
+                fileOutputStream.write(buf, 0, nBytes);
+                buf = new byte[DATA_SIZE];
             }
+
+            System.out.println("Received file '" + filename + "' from " + socket.getInetAddress().getHostAddress() + ":" + socket.getPort());
+
+            fileOutputStream.close();
+            in.close();
+            out.close();
+            socket.close();
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
     }
 
 }
